+++ plugins/main/public/components/common/modules/modules-defaults.tsx
@@ -22,6 +22,7 @@ import {
 } from '../wazuh-discover/wz-discover';
 import { CaseManagementTab } from '../document-details/case-management';
 import { threatHuntingColumns } from '../../overview/threat-hunting/events/threat-hunting-columns';
+import { ThreatHuntingCases } from '../../overview';
 import { vulnerabilitiesColumns } from '../../overview/vulnerabilities/events/vulnerabilities-columns';
 import { dockerColumns } from '../../overview/docker/events/docker-columns';
 import { googleCloudColumns } from '../../overview/google-cloud/events/google-cloud-columns';
@@ -172,6 +173,12 @@ export const ModulesDefaults = {
           WAZUH_SAMPLE_ALERTS_CATEGORY_THREAT_DETECTION,
         ],
       }),
+      {
+        id: 'cases',
+        name: 'Cases',
+        buttons: [ButtonExploreAgent],
+        component: ThreatHuntingCases,
+      },
     ],
     availableFor: ['manager', 'agent'],
   },