+++ plugins/main/public/components/overview/threat-hunting/cases/table-columns.ts
@@ -1,0 +1,9 @@
+import { commonColumns } from '../../common/data-grid-columns';
+
+export default [
+  commonColumns.timestamp,
+  commonColumns['wazuh.agent.name'],
+  { id: 'wazuh.case.status' },
+  { id: 'wazuh.case.tags' },
+  { id: 'wazuh.case.comment' },
+];